feat: suppress safe shadowing within fun binders
This PR modifies pretty printing of `fun` binders, suppressing the safe shadowing feature among the binders in the same `fun`. For example, rather than pretty printing as `fun x x => 0`, we now see `fun x x_1 => 0`. The calculation is done per `fun`, so for example `fun x => id fun x => 0` pretty prints as-is, taking advantage of safe shadowing.

The motivation for this change is that many users have reported that safe shadowing within the same `fun` is confusing.