feat: include `:=` in the atomic part of tactic configuration items
This PR modifies the syntax for tactic configurations. Previously just `(ident` would commit to tactic configuration item parsing, but now it needs to be `(ident :=`. This enables reliably using tactic configurations before the `term` category. For example, given `syntax "my_tac" optConfig term : tactic`, it used to be that `my_tac (x + y)` would have an error on `+` with "expected `:=`", but now it parses the term.

An additional rationale is that these are like named arguments; (1) terms can't begin with named arguments and (2) `Parser.Term.namedArgument` indeed already includes `:=` in the atomic part.