feat: improvements to Verso docstrings
This PR implements module docstrings in Verso syntax, as well as
adding a number of improvements and fixes to Verso docstrings in
general. In particular, they now have language server support and are
parsed at parse time rather than elaboration time, so the snapshot's
syntax tree includes the parsed documentation.