fix: use `quick_exit` in the language server
We have previously observed deadlocking calls to `exit` inside of the
language server and deemed them irrelevant. However, child processes of
these deadlocking exiting processes can continue to consume a large
amount of CPU as they try to compile a library etc. Hence, this PR
switches to the MT safe `quick_exit` inside of the language server,
in order to ensure the server finishes when it is told to.