chore: ensure `pass` refers to `SpecResult.pass` in GuardMsgs
This PR has a tiny change, it just adds a `.` in front of `pass`. Previously, the match arm read `| pass => ...`. Presumably, `pass` was intended to mean `SpecResult.pass`, but, this isn't in scope, so instead `pass` here is a catch-all variable. By adding a dot, we ensure we actually refer to the constant.

Note that this was the last case in the pattern-match, and since all other constructors were correctly referenced, the only case that went to the fallback was `SpecResult.pass`, so the code did the right thing. Still, by fixing this, we prevent a surprise in the event that a new `SpecResult` constructor is added.