perf: experiment: abstract recursors in recursor's RHSs
This PR experiments with a slightly different representation of a
recursor's rules. Instead of having a rules
```
fun α motive rec cons {n} x xs => cons x xs (Vec.rec α motive rec cons n xs)
```
we leave a hole for `Vec.rec α motive rec` and abstract over this
```
fun α motive rec nil cons {n} x xs => cons x xs (rec n xs)
```

This could be beneficial if the recursor has many arguments (e.g. many
minor premises from an inductive with many constructors). Note that the
subexpression `Vec.rec α motive rec` is exactly a subterm of the
expression we are reducing. So instead of substituting every single
argument, we can substitute this subexpression as a whole, increasing
sharing and reducing the cost.

It also simplifies the kernel code for nested inductives a bit, as
recursor names don't have to be changed in the rules `rhs`, only in the
list of mutual recursors.

This refactoring is also in anticipation of maybe treating more
functions as primitively recursive, instead of always going all the way
to `.rec`.