fix: Verso docstring semantic highlighting fixes
This PR re-enables semantic tokens for Verso docstrings, after a prior
change accidentally disabled them. It also adds a test to prevent this
from happening again.

In the process, it became clear that there was a bug that led to
overlapping but not identical tokens, but the code had previously
assumed that this couldn't happen at the delta-encoding step. So this
PR additionally replaces the removal of duplicate tokens with
priority-based handling of overlapping tokens.