feat: remove support for reducible well-founded recursion
This PR removes support for reducible well-founded recursion, a
Breaking Change. Using `@[semireducible]` on a definition by
well-founded recursion prints a warning that this is no longer effective.

With the upcoming module system, proofs are often not available. With
this change, we remove a fringe use case hat may require proofs.

At least for now, direct use of `WellFounded.fix` are not affected.