feat: mkAuxTheorem to not abstract over aux decl
This PR teaches `mkAuxTheorem` to not abstract over fvars that are
marked as “aux declaration”, i.e. a placeholder for recursive call, but
instead abstract over the whole application headed by that fvar. This
way the recursive call is still visible for termination checking.

Before, termination checking would undo the abstraction of proofs here,
but this is roundabout and unreliable. So this fixes #10721.

Not abstracting over recursive calls here also helps with matchers
abstracting over such calls, so this fixes #2102.