fix: hovers and docstrings for (co)inductive types
This PR fixes a regression introduced by #10307, where hovering the
name of an inductive type or constructor in its own declaration didn't
show the docstring. In the process, a bug in docstring handling for
coinductive types was discovered and also fixed. Tests are added to
prevent the regression from repeating in the future.