fix: avoid unnecessary branching in match compilation
This PR improves match compilation: Branch on variables in the order
suggested by the first remaining alternative, and do not branch when the
first remaining alternative does not require it. This fixes #10749.

(For now this is an experiment to get familiar with the code and the whole
problem domain. It is likely overly naive.)